requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
deep-sort-realtime>=1.3.2
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
click>=8.1.0
filterpy>=1.4.5
scipy>=1.10.0
